backend/requirements.txt
fastapi[standard]
uvicorn
pydantic
firebase-admin
google-generativeai
google-genai
google-cloud-logging
requests
beautifulsoup4
